[#52] Refactor: HiService 분리 및 기존 코드 리팩토링
