+++ src/main/java/com/gdg/z_meet/domain/meeting/service/MeetingQueryServiceImpl.java
@@ -42,7 +42,6 @@ public class MeetingQueryServiceImpl implements MeetingQueryService {
     private final UserProfileRepository userProfileRepository;
     private final TeamRepository teamRepository;
     private final UserTeamRepository userTeamRepository;
-    private final HiRepository hiRepository;
 
     @Override
     @Transactional(readOnly = true)
@@ -138,86 +137,22 @@ public class MeetingQueryServiceImpl implements MeetingQueryService {
                 .build();
     }
 
-    @Override
-    @Transactional()
-    public void sendHi(MeetingRequestDTO.hiDto hiDto){
-        List<Long> teamIds = Arrays.asList(hiDto.getFromId(), hiDto.getToId());
-        List<Team> teams = teamRepository.findByIdIn(teamIds);
 
-        if (teams.size() != 2) {
-            throw new BusinessException(Code.TEAM_NOT_FOUND); // 모든 팀을 못 찾은 경우
-        }
-
-        Team from = null;
-        Team to = null;
-
-        // 팀 순서 확인 후 할당
-        if (teams.get(0).getId().equals(hiDto.getFromId())) {
-            from = teams.get(0); // 보내는 팀
-            to = teams.get(1);   // 받는 팀
-        } else {
-            from = teams.get(1); // 보내는 팀
-            to = teams.get(0);   // 받는 팀
-        }
-
-
-        if(from.getTeamType()!=to.getTeamType()) throw new BusinessException(Code.TEAM_TYPE_MISMATCH);
-        if(from.getGender()==to.getGender()) throw new BusinessException(Code.SAME_GENDER);
-        if(hiRepository.existsByFromAndTo(from,to)) throw new BusinessException(Code.HI_DUPLICATION);
-
-        from.decreaseHi(); // 하이 갯수 차감
-
-        Hi hi = Hi.builder()
-                .hiStatus(HiStatus.NONE)
-                .from(from)
-                .to(to)
-                .build();
-        hiRepository.save(hi);
-    }
-
-    @Override
-    @Transactional
-    public void refuseHi(MeetingRequestDTO.hiDto hiDto) {
-        List<Long> teamIds = Arrays.asList(hiDto.getFromId(), hiDto.getToId());
-        List<Team> teams = teamRepository.findByIdIn(teamIds);
-
-        if (teams.size() != 2) {
-            throw new BusinessException(Code.TEAM_NOT_FOUND); // 모든 팀을 못 찾은 경우
-        }
-
-        Team from = null;
-        Team to = null;
-
-        // 팀 순서 확인 후 할당
-        if (teams.get(0).getId().equals(hiDto.getFromId())) {
-            from = teams.get(0); // 보내는 팀
-            to = teams.get(1);   // 받는 팀
-        } else {
-            from = teams.get(1); // 보내는 팀
-            to = teams.get(0);   // 받는 팀
-        }
-
-        Hi hi = hiRepository.findByFromAndTo(from,to);
-        if(hi==null) throw new BusinessException(Code.HI_NOT_FOUND);
-        hi.changeStatus(HiStatus.REFUSE);
-        hiRepository.save(hi);
-    }
-
-    private Map<Long, List<String>> collectEmoji(List<Team> teamList) {
+    public Map<Long, List<String>> collectEmoji(List<Team> teamList) {
 
         return collectTeamInfo(teamList,
                 userTeam -> userTeam.getUser().getUserProfile().getEmoji(),
                 false);
     }
 
-    private Map<Long, List<String>> collectMajor(List<Team> teamList) {
+    public Map<Long, List<String>> collectMajor(List<Team> teamList) {
 
         return collectTeamInfo(teamList,
                 userTeam -> String.valueOf(userTeam.getUser().getUserProfile().getMajor()),
                 true);
     }
 
-    private Map<Long, Double> collectAge(List<Team> teamList) {
+    public Map<Long, Double> collectAge(List<Team> teamList) {
 
         return teamList.stream().collect(Collectors.toMap(
                 Team::getId, team -> userTeamRepository.findByTeamId(team.getId()).stream()
@@ -227,7 +162,7 @@ public class MeetingQueryServiceImpl implements MeetingQueryService {
         ));
     }
 
-    private Map<Long, List<String>> collectMusic(List<Team> teamList) {
+    public Map<Long, List<String>> collectMusic(List<Team> teamList) {
 
         return collectTeamInfo(teamList,
                 userTeam -> String.valueOf(userTeam.getUser().getUserProfile().getMusic()),
@@ -260,97 +195,6 @@ public class MeetingQueryServiceImpl implements MeetingQueryService {
         if (teamType == TeamType.THREE_TO_THREE && userCount != 3) {
             throw new BusinessException(Code.TEAM_TYPE_MISMATCH);
         }
-    }
-
-    @Override
-    @Transactional
-    public List<MeetingResponseDTO.hiListDto> checkHiList(Long teamId, String action){
-        List<Hi> hiList;
-        List<Long> teamIds;
-        if(action.equals("Receive")) {
-            hiList = hiRepository.findRecevieHiList(teamId);
-            teamIds = hiList.stream()
-                    .map(hi -> hi.getFrom().getId()) // 보낸 팀의 id
-                    .collect(Collectors.toList());
-        }
-        else{
-            hiList = hiRepository.findSendHiList(teamId);
-            teamIds = hiList.stream()
-                    .map(hi -> hi.getTo().getId())
-                    .collect(Collectors.toList());
-        }
-
-        List<Team> teamList = teamRepository.findByIdIn(teamIds);
-        System.out.println("TeamList: "+teamList);
-
-        Map<Long, List<String>> emojiList = collectEmoji(teamList);
-        Map<Long, List<String>> majorList = collectMajor(teamList);
-        Map<Long, Double> age = collectAge(teamList);
-        Map<Long, List<String>> musicList = collectMusic(teamList);
-
-
-        // 여러 개의 hiListDto 생성
-        List<MeetingResponseDTO.hiListDto> hiListDtos = new ArrayList<>();
-        for (Hi hi : hiList) {
-            if(hi.getHiStatus()!=HiStatus.NONE && action.equals("Receive")) continue;
-
-            Team team = action.equals("Receive") ? hi.getFrom() : hi.getTo();
-
-            // UserProfileDto 생성 (null 방지)
-            String major = String.join(", ", majorList.getOrDefault(team.getId(), Collections.emptyList()));
-            String emoji = String.join(", ", emojiList.getOrDefault(team.getId(), Collections.emptyList()));
-            String music = String.join(", ", musicList.getOrDefault(team.getId(), Collections.emptyList()));
-
-            // 각 팀에 대해 UserProfileDto 만들기
-            List<MeetingResponseDTO.hiListDto.UserProfileDto> userProfileDtos = new ArrayList<>();
-            MeetingResponseDTO.hiListDto.UserProfileDto userProfileDto = MeetingResponseDTO.hiListDto.UserProfileDto.builder()
-                    .major(major)
-                    .emoji(emoji)
-                    .music(music)
-                    .build();
-            userProfileDtos.add(userProfileDto);
-
-            LocalDateTime sentTime = hi.getCreatedAt(); // Hi 생성 시간
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MM-dd HH:mm");
-            String dateTime = String.format("%s 전송", sentTime.format(formatter));
-
-            if(action.equals("Receive")){
-                LocalDateTime now = LocalDateTime.now(); // 현재 시간
-                Duration duration = Duration.between(sentTime, now);
-
-                long totalMinutesElapsed = duration.toMinutes(); // 보낸 후 총 경과된 분
-                long totalMinutesRemaining = (5 * 60) - totalMinutesElapsed; // 5시간(300분) 기준으로 남은 분 계산
-
-                if (totalMinutesRemaining <= 0) {
-                    hi.changeStatus(HiStatus.REFUSE);
-                    hiRepository.save(hi);
-                    continue;
-                }
-
-                long remainingHours = totalMinutesRemaining / 60; // 남은 시간을 60으로 나눠서 시간 계산
-                long remainingMinutes = totalMinutesRemaining % 60; // 나머지 분 계산
-
-                dateTime = String.format("%d시간 %d분 남음", remainingHours, remainingMinutes);
-
-               if(remainingHours<=0 || remainingMinutes<=0){
-                   hi.changeStatus(HiStatus.REFUSE);
-                   hiRepository.save(hi);
-                   continue;
-               }
-           }
-
-            // 하나의 hiListDto 생성
-            MeetingResponseDTO.hiListDto hiDto = MeetingResponseDTO.hiListDto.builder()
-                    .teamName(team.getName())
-                    .teamList(userProfileDtos)
-                    .age(Math.round(age.get(team.getId()) * 10.0) / 10.0)
-                    .dateTime(dateTime)
-                    .build();
-
-            hiListDtos.add(hiDto); // hiListDto를 리스트에 추가
-        }
-
-        return hiListDtos;
     }
